PCP-215: @sidenav @content 슬롯 제거 (병렬 라우트 해제 및 롤백)
