--- src/app/team/[teamId]/@content/page.tsx
@@ -1,8 +1,0 @@
-import { redirect } from 'next/navigation';
-
-const page = async ({ params }: { params: Promise<{ teamId: string }> }) => {
-  const { teamId } = await params;
-  redirect(`/team/${teamId}/dashboard`);
-};
-
-export default page;
