+++ src/app/team/[teamId]/manage/setting/page.tsx
